fixed overflow bug of sys:generate_collective, int tmp -> uint64_t tmp 	modified:   astra-sim/system/Sys.cc

+++ astra-sim/system/Sys.cc
@@ -766,7 +766,7 @@ DataSet* Sys::generate_collective(
   uint64_t chunk_size = determine_chunk_size(size, collective_type);
   uint64_t recommended_chunk_size = chunk_size;
   int streams = ceil(((double)size) / chunk_size);
-  int tmp;
+  int64_t tmp;
   DataSet* dataset = new DataSet(streams);
   int pri = get_priority(explicit_priority);
   int count = 0;
